Make window resizable again.
Note: I can't simple place SDL_WINDOW_RESIZABLE in _fullscreen_modes[] because SDL ignores this flag in SDL_SetWindowFullscreen and if the user started the game in fullscreen and switched to window, it wouldn't be resizable.
The way I did it is the only way I know to make the window resizable regardless of whether the game was first started in window or in fullscreen.