Updated lots of ride descriptions and other things
'air time' is now 'hang tijd', we might just want to keep calling it 'air time', I'm not certain.

Changed autorondrit to rit in een karretje, which is what I remembered it as from my youth. Smalspoorweg also got changed to miniatuur spoorbaan, as most theme parks would refer to them as such.

Renamed the Heartline-twisteractbaan to Heartline-kurkentrekkerachtbaan. Twister = kurkentrekker.

Stoelrot. changed to Stoeldraaing (I'm not sure if it fits, otherwise revert to Stoelrot.)

Also changed a whole bunch of other things down the road. This is fun!