Here a few more improvements, still do not merge it!
What do the others think of it? 
STR_0569: we zijn niet bij de Flodders thuis! 
STR_0586: not sure about this change. Also the sentence itself is a bit too long and odd. Does anyone have suggestions to improve this sentence or shall we leave it as it is? 
STR_0908, STR_0910-STR_0913: Where is this string for (expecially STR_0908) and what does it mean?! So, what does roll means in Dutch? Rollen or looping or something? I think we have to change roll into something else...