And again a few improvements...
STR_1015: Is OpenRCT2 able to display these special characters?
STR_1080: Should we translate the names of the movies?
STR_1093: Should we translate this one and when yes, into what?
STR_1099: Shoud the second "op" be replaced here by "van" or something?
STR_1174: Is there no better translation of the word "Banner"? I have never heard of the word "Banier."
STR_1357: To what should we translate "holes"?
STR_1365: Should we translate this one, when yes, into what?
STR_1370: Into what should we translate Barrel roll? STR_1378: And what to do with Heartline roll?
STR_1673: Does someone know a better translation for Spinning Control Toggle Track, as Draaistop looks a bit odd now.
STR_1760: Shouldn't this be translated to Invulmodus or Vervangmodus or something? This is now a strange/odd translation