Fix #958.
Issue was caused by an old bit of code that had been added in the force weather cheat to keep it sunny for longer before realising how the climate operated correctly. Instead now it sets the weather to the desired value and starts a normal amount of waiting time before returning to the normal weather cycle.