Fix scenario bugs.
Number of guests and currency were accidentaly being truncated to bytes. This was causing display issues when creating custom scenarios in OpenRCT2.
There was a small bug in research loading that was causing a crash due to exiting of a loop to early. As this was after a call to random this would only occur occasionally.