Fix crash on loading certain sv6 files
Sub_684AC3 was incorrectly being called on .sv6 load. Not completely sure on the function of 684ac3 but I think it marks anything that should be researched as researched. Since .sv6 files have already used the research list it shouldn't be reset. I'm unsure why its crashing though. Perhaps this save was used with an old version of the scenario creator and the list is corrupted.