Grant MMX to non-windows platforms
As there is still of un-reversed code, especially the
peep/rides/terrain drawing, a likely potential candidate for using MMX,
these instruction can potentially be useful in making things
speedier even when not using Windows.

I tested it by running a park for a while and saw no regression. I
haven't went in as deep as to actually verify if MMX is used or not.