Fix #2173. Entrance fee too high can no longer be a news item if no money is on.
This is kind of an original bug because the game doesn't check normally to see if money is on during this. I didn't want to reset the entrance fee as then it would be annoying switching between the two states. The peeps don't pay if no money is on so nothing is lost.