Fix some left-shifts
An explicit `1` is considered signed by compiler, if `int` is 32 bit
long, a left shift `1 << 31` is undefined. Since the type of result is
the type of left shift operand, make sure we shift unsigned int so that
the behaviour is defined correctly.

Also one fix in peep code, where it would use too big value for
`item_extra_flags`.