Handle touch drag on viewports, fixes #2091
This commit makes it possible to drag a viewport around with touch.
The user can drag if they start without a ride, shop, etc under
their touch.  This reuses the code of the right mouse button viewport
movement, although adding a new flag to indicate touch behavious
(as well as not relying on moving the cursor).