OS X job for Travis CI
This adds automated builds for OS X to our CI infrastructure.

XCode's `xcodebuild` tool is quite spammy by default, so I used a
prettifier for its output. Since this could potentially cover errors
from XCode due to piping of commands, I enabled `pipefail` option which
would cause all of pipe construct fail, should any of its commands exit
with non-zero code.

The created app bundle is packaged into dmg and uploaded to transfer.sh,
if the build was successful. To the best of my knowledge, this is *the*
way of packaging and distributing OS X applications.