Skip "Default sound device" on Linux
Sound on Linux is handled by PulseAudio, which creates a sink and offers
a central place to control which sources are distributed how and where.

Applications in general are expected to only offer volume controls.

Tested with 0 A.D., SuperTuxKart and a few media players, none of which
offers selecting sink directly.

As per https://en.wikipedia.org/wiki/PulseAudio#Adoption it should
covert vast majority of uses.