Disable Cluster build in MP Client
Due to the frequent abuse of the cluster build button, it feels only right to come up with some sort of temporary damage control until someone can modify the permissions system to allow the cluster build not to pass through. This solution may be bypassable, but it is meant only to provide a temporary workaround. People that are out to vandalise will likely not take the extra effort to vandalise if they do not know how to bypass it, and manually clicking in objects would not be as convenient and likely will prevent them from trying, or at least get a chance to find out who is vandalising the park before they really damage it. I feel disabling the button in multiplayer clients would be worth it, the only downside is things such as forests and schrubbery will require manual placement, but the compromise feels worth it.