Handle touch drag on viewports, fixes #2091
This commit makes it possible to drag a viewport around with touch.
This reuses the code of the right mouse button viewport
movement, although adding a new flag to indicate touch behavious

The user can always start a drag, however the drag is teated as a
click if the user does not move more than 8px.
(as well as not relying on moving the cursor).