require special permission for cluster building
Because I do not know too much about game_do_command yet, i did not want
to modify it at the expense of introducing bugs. This currently serves
no purpose other than to allow a game command function, as the
functionality uses network_can_do_action instead. It's a little hackish,
but it gets the job done for right now.