CLUSTER: Fixed some stuff and improved server-side prevention
Crash tested it and found a coupleof problems that I fixed, I switched
from using an empty gamecommand to -2 like people suggested, I also
added some additional server-side prevention incase someone modifies the
code to allow it on their own client

Because there can now provide server-side-only prevention (rather than
requiring a user to have the latest version), I bumped the stream
version back down to 4 in the meantime. When it is appropriate, we can
change the stream version to 5.