Cast peep's arguments properly
Left-hand side of shift operator determines what's the resultant size of
whole expression. Consider this:

    uint16 p = 65535;
    uint32 q = p << 16;

`q` will be an undefined result, as it will not fit in `uint16` type
anymore, and compiler can optimise it away, unless you cast `p` to be
something which can hold such value, in this case `uint32`.