remove VS filters
The project filters for VS is just becoming an inconvenience and always ends up going out of sync with the src folder structure. Developers should use 'Show All Files' instead which displays the source files by their actual folder structure and other files in the repository that aren't included in the vcxproj.