Customizable Kick Reason
This commit provides the ability to customize the message shown when
kicking by selecting from a few predefined options, or adding your own
option. I have rebased and squashed all the commits to prevent this pull
request from clogging up the OpenRCT2 history (the majority of the commits
were meren't very maintained and often just pushed so i can jump to a
seperate branch or something). I have thuroughly tested this, though
additional testing is always welcome before it is merged. if you test it,
make sure to test it in your own server, as it renders the previous kick
method obsolete and incompatible. I feel it is in
a well enough state that i can call it "finished".