Allow corner tiles when using mountain tool
The tile selection was limited to full tiles when using the mountain tool on a single tile. The game now allows selecting and raising/lowering corner tiles and the smoothing algorithm is modified to support that. The cost is the same as the old single-tile mountain tool. A bug where the mountain tool sometimes creates sharp edges has also been fixed.