Fix the peep path finding
In the heuristic search, tiles in wide path were always being handled as junctions.  Changed to handle junctions according to the path thinning.

There was no search loop detection.  Added loop detection for returning to the starting location and any of the junctions visited in the current search path.

Various other optimisations added.