Fix #4231. Ride colour issue caused by overlapping vars.
By using a union instead of a struct var_494 and var_496 where ontop of the same location this meant var_496 when written to would overwrite ride_colour that was also in a union with var_494/496. This in turn would cause the wrong ride_colour to be overwritten when using the colour picker.