Fix issues with sync adjacent stations
Change try_add_synchronised_station() so that adjacent stations are always in the _synchronisedVehicles list; those without a vehicle to sync have vehicle_id == SPRITE_INDEX_NULL.

Change vehicle_can_depart_synchronised() so that adjacent stations are searched for on both sides of the tile; reverse the logic of the return values to make sense with respect to the function name (all except for the case when there are no adjacent stations, which already made sense and was consequently flipped in relation to all other returns; numerous minor fixes to the checks on the departure conditions/exceptions.

In vehicle_update_waiting_to_depart() flip the logic on the call to vehicle_can_depart_synchronised() according with the above change to its return values.

Fixes issues: #4178 and #4182