Ignore ride queues when determining 'thin' junctions.
Why? In prinicple, a queue connected to a path should not be seen as causing a junction in that path.  Rather a ride queue should be treated equivalently to ride entrances/exits. For the path finding search this means that maxNumJunctions are not consumed when walking past ride queues (unless there is a legitimate path junction there).

Remove the unused arguement 'dist' from get_ride_queue_end().