Fix broken zoomed textures and palettes
Zoomed textures were failing to pass the image flags to the zoomed version of the texture. This would cause corrupted data to be passed as a tertiary colour to the palette function.
Also due to not recording the image flags in the palette hash map a tertiary colour of zero would be ignored and unremaped pink colour would be shown