Refactor naming of paint global vars
Removed EDF81C from none viewport files this previously was used to pass image flags to the sprite renderer.
gUnkF1A4CC replaced with address reference to paint struct array to allow for non rct builds
Renamed _paint_structs to _paint_struct_quadrants. This array is used for working out which paint struct is at which quadrant and is not the main paint struct array