Fix #3062: Backtick ` opens console when inputting text into chat or other things
For some reason SDL_IsTextInputActive() returns SDL_TRUE before any call to SDL_StartTextInput(). This may be an SDL bug, but I've worked around it for this feature by also checking if there is a valid gTextInput.buffer. I've made an exception for the console itself, where the backtick should probably continue to close it.