Fix #3355: Implement loading of parks from URLs
The help text for the command line options already referenced the
possibility of opening a saved park directly from a URL, but this was
not yet implemented. This commit changes all path handling for command
line options to accept both local paths and URLs.

If a URL is specified instead of a local path, the program will download
the file to the operating system's temp directory. It will then proceed
to load this file just like local files would be loaded.

The program will try to derive the extension of the temp file from the
original URL and defaults to sv6 (a save file) if it is unable to do so.