Disable MSVC 2017 RC on AppVeyor
I don't recall a build that would fail on MSVC15, but not on MSVC14.
Testing with both of them provides little benefit, while causing huge
delays in builds on AppVeyor.

In due time this can be re-enabled or AppVeyor setup can be moved fully
to MSVC15 when it becomes released.