Add support for inserting corrupt elements in MP
This PR adds three new files, one to expose some of the funcitons for the
tile inspector window, and two completely new ones for the tile inspector
logic, which handles what happens.

This commit adds the logic for inserting corrupt elements, while keeping
the list on both server and client correct (e.g. client has the same tile
selected, and keeps the same element in focus when the tiles gets changed)