remove excess generate_new_guest, make park_generate_new_guest public
I noticed there is an extra generate_new_guest which calls park_generate_new_guest. This can be removed in favor of just making park_generate_new_guest public.

This also allows potential in the future for customized trams, such as a group of angry people arriving to trash the park or something like that.