Remove duplicate ride penalty for closed rides
The game reduces the 'value' of a ride, which decides how much guests will pay for it, when multiple rides of the same type exists within a park. Currently, the rides' status is not taken into account.

This PR ensures a ride with status "closed" will no longer attract the penalty. Only open or broken rides will do so.

In other words, this PR removes ill-deserved punishment for players who:

- Use (partial) rides as scenery
- Build replacement rides while keeping the originals open

Also refactored the code slightly to decrease average-case runtime.