Prevent peeps walking up and back down ride queues
This bug is caused by inconsistent use of peep->interactionRideIndex. The primary use of this value is to record when a peep has decided to go on the ride at the queue entrance so they remember that decision when entering each queue path tile - without this they would re-decide whether to go on the ride each time they walk into a new ride queue tile.

This is also important for making the peeps walk up and back out of the queue when the ride is closed.

This value is also set when a peep decides not to go on a ride (though the reason why is not clear to me), but is not reset/changed on interaction with other tile types (e.g. shops, no entry signs, etc) causing peeps who bounce back on a ride queue entrance to walk up and down the queue.

Fixes #3499

Requires new network version.