Allow mechanics to walk over wide tiles on the edge of the patrol zone
This is a better alternative to #5424 and #5429.

With this fix, the heuristic allow mechanics to walk onto tiles on the patrol edge with the wide flag set, but does not otherwise change the wide flag for the rest of the heuristic which was causing serious problems in some cases in #5424 and #5429.