Also ignore wide flag on current path when attempting to walk onto a wide path.
With the previous commit, wide path flags are only checked on the path the peep is walking onto. For staff with a patrol area, the ignored wide flags unintentionally permits staff to walk onto other wide paths. In some circumstances, when the player is editing the path to the peep destination and the path to the destination is temporarily disconnected, this results in the heuristic not failing as intended (which resets the pathfind history) and the peep gets stuck. When the player reconnects the path to the destination, the peep will remain stuck if junctions to the destination are in the pathfind history.