Fix #5890. Primary and secondary colours now work in zoomed opengl. (#6109)
Mistake made due to a missunderstanding of how the software renderer works. At this point in the software renderer the primary and secondary colours are already calculated into the palette so can be ignored. In opengl the colours still need to be passed to the command queue. Fixes #4715 as well.

Update changelog