Fix Guest List summary not updating after a ride rename
The Guest List summary screen only updates every several seconds, so it does not affect the regular Guest list that updates quickly. If the Guest List summary page is open while renaming a ride, the ride name string will become blank for any actions/thoughts until the auto refresh kicks in. This fix forces a refresh on the summary page (if opened) when a ride is renamed. This works in multiplayer as well.