Adding gridlines display toggle shortcut.
This feature existed in the original RCT1/2 games but was not present in
the latest development commit. There remains work to be done with how
this shortcut interacts with both the saved preference for gridlines
display as well as the construction tools, so for the time being it
simply toggles the gridlines display depending on its current state.

The original shortcut for this key was '7', placing it between the
"Invisible people toggle" (6) shortcut and the "Height marks on land
toggle" (8). This commit preserves as much as possible, using
placeholder comments that signify the relevant values which this
feature actually uses where this is not possible.